Fix pylint and pydocstyle warnings
